JobPostingParser.css 면접리스트 디자인 수정
