+++ src/components/jobposting/JobPostingParser.jsx
@@ -33,24 +33,15 @@ const JobPostingParser = () => {
     }
   };
 
-  // ✅ 기업명 다듬기 (대괄호나 여백 제거)
   const refineCompanyName = (name) => {
     if (!name) return "기업명 미상";
-    // "[원티드랩]" 처럼 오는 경우 대괄호를 제거합니다.
     return name.replace(/\[|\]/g, "").trim();
   };
 
-  // ✅ 직무명 다듬기 (지저분한 뒷부분만 날리고, 앞부분의 [프론트엔드] 같은 정보는 살림)
   const refineJobTitle = (title) => {
     if (!title) return "직무 미상";
-    
-    // "채용 공고 | 원티드" 같이 공통적으로 붙는 지저분한 꼬리표만 정교하게 자릅니다.
-    // split() 대신 정규식으로 '채용'이라는 단어 뒤의 모든 것을 날리되, 
-    // "채용 연계형 인턴" 같은 경우는 날아가지 않도록 보통 맨 뒤에 붙는 패턴만 날립니다.
     let refined = title.replace(/(채용\s*공고\s*\|).*$/g, "").trim();
     refined = refined.replace(/\|\s*원티드.*$/g, "").trim();
-    
-    // 만약 파싱된 제목이 "프론트엔드 개발자 [신규서비스]" 라면 그대로 살려둡니다.
     return refined || title; 
   };
 
@@ -63,16 +54,6 @@ const JobPostingParser = () => {
         ))}
       </div>
     );
-  };
-
-  const renderListItems = (data, emptyMessage) => {
-    if (!data) return null;
-    if (Array.isArray(data)) {
-      return data.length > 0 ? (
-        data.map((item, idx) => <li key={idx}>{item}</li>)
-      ) : null;
-    }
-    return <li>{String(data)}</li>;
   };
 
   const renderTextWithLineBreaks = (text) => {
@@ -115,7 +96,6 @@ const JobPostingParser = () => {
       {parsedData && (
         <div className="job-parser-result-box">
           <div className="job-parser-header">
-            {/* ✅ 하드코딩된 대괄호 [] 를 제거하고, refineCompanyName 함수 적용 */}
             <span className="job-parser-company-badge">
               {refineCompanyName(parsedData.companyName)}
             </span>
@@ -124,7 +104,6 @@ const JobPostingParser = () => {
             </h4>
           </div>
 
-          {/* ... (이하 렌더링 코드는 이전과 완전히 동일) ... */}
           {parsedData.mainTasks && parsedData.mainTasks !== "정보 없음" && parsedData.mainTasks !== "없음" && (
             <div className="job-parser-info-section">
               <strong>🔹 주요 업무</strong>
@@ -160,14 +139,24 @@ const JobPostingParser = () => {
             </div>
           )}
 
+          {/* ✅ 면접 예상 질문 영역을 화려하게 포장 */}
           {Array.isArray(parsedData.expectedQuestions) && parsedData.expectedQuestions.length > 0 && (
-            <div className="job-parser-info-section">
-              <strong>🎯 면접 예상 질문</strong>
-              <ul className="job-parser-question-list">
-                {renderListItems(parsedData.expectedQuestions, "예상 질문을 생성하지 못했습니다.")}
-              </ul>
+            <div className="job-parser-questions-section">
+              <h4 className="job-parser-questions-title">🎯 AI 면접 예상 질문</h4>
+              <p className="job-parser-questions-desc">
+                지원하신 직무와 우대사항을 바탕으로 추출한 핵심 면접 질문입니다.
+              </p>
+              <div className="job-parser-questions-list">
+                {parsedData.expectedQuestions.map((q, idx) => (
+                  <div key={idx} className="job-parser-question-card">
+                    <div className="job-parser-question-badge">Q{idx + 1}</div>
+                    <div className="job-parser-question-text">{q}</div>
+                  </div>
+                ))}
+              </div>
             </div>
           )}
+          
         </div>
       )}
     </div>
